Develop T3b-2: the masks core loses its widgets
The function move T3b-1's header boundary made mechanical: 91 functions, 3173
lines, out of masks.c into masks_gui.c -- hit testing, event dispatch,
hover/cursor, form_gui lifecycle, creation flows, drawing, menus, the
interaction engine. Classified by transitive closure from each header's
declarations (the exif.cc recipe: bare-identifier references, byte-conserving
cut, every function placed exactly once), with placement decided by declaring
header and seven shared statics de-static'd into masks/masks_functions.h.

Three functions needed judgement rather than mechanics. dt_masks_set_edit_mode
is engine-callable (dev_history restores it on undo) but pokes the blending
panel's toggle -- it moves to the GUI half whole, its masks.h declaration
unchanged, exactly like dt_dev_history_gui_update. _change_opacity carried a
toast inside the value-mutation path -- the trap the survey flagged -- and
dt_masks_form_change_opacity is called only from shape scroll handlers; both are
interaction machinery and move. The last dt_control_log (mask version mismatch
on DB read) goes through dt_history_message(), the seam that already owns
persistence complaints.

masks.c ends at 1473 lines with ZERO upward includes and ZERO toolkit symbols:
group data ops, refcounting, persistence, legacy params, gravity centers, hashes
and the rasterisation dispatchers. Seven GUI includes dropped, plus a
control/signal.h whose raises had left with the moved code.

Release, Debug (-Werror), nofeatures all build.
